LHRS optics added compare_DB script to compare MEs form two DBs
Added compare_DBs.C (and header compare_DBs.h). This script is
designed to take in two optics DBs and compare their Matrix Elements
for each order (those MEs which exist in both). The type of ME
compared can be selected (T,Y,P or D).

The script produces (for each order of ME existent) a plot of both ME
values, a plot of the absolute proportional difference and a plot of
the absolute difference.

Ran in root by:
.x compare_DBs("DB_name_1","DB_name_2",element)
: where "DB_name_1" and "DB_name_2" are TStrings which select the DBs
to be compared and element is an Int_t (int) which can be (T,Y,P or D)
used to select the type of element beign compared.